# Making encode_col loads faster