Benchmark looks like this

```
Settings:
  duration:      1.0 s

## EncodeColBench
[16:12:47] 1/2: mine
[16:12:52] 2/2: original

Finished in 7.01 seconds

## EncodeColBench
benchmark iterations   average time 
mine        10000000   0.46 µs/op
original      100000   17.00 µs/op
```