@mikebaldry 2017-07-13
My spreadsheet contains lots of data, 170 sheets, each with 366 columns and ~80 rows. So this makes a difference :)